Removed explicit requests to load GLcore and bitmap modules.
GLcore caused the X server to segfault when running any OpenGL application
with indirect rendering because of symbol clash with *_dri.so (loaded
by AIGLX). OpenGL is still available, because the glx module automatically
loads either the relevant DRI module, or, if it is impossible, GLcore.

The bitmap module is not needed because the freetype module handles the two
builtin bitmap fonts well.